Apply more suggestions
* Rephrase `--backup-path` and `--no-backup` docs
* Add empty lines to `rotate_saved_backup_files()`
* Remove redundant `#[clap(conflicts_with)]`
* Make `".abfts"` a constant instead of a variable
* Factor backup-related items out into a separate module
